+++ components/function-controller/internal/controllers/serverless/function_controller_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"math/rand"
-	"testing"
 	"time"
 
 	"github.com/onsi/ginkgo"
@@ -12,7 +11,6 @@ import (
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/tools/record"
@@ -200,126 +198,3 @@ func newFixFunction(namespace, name string) *serverlessv1alpha1.Function {
 	}
 }
 
-func Test_getNewestGeneration(t *testing.T) {
-	type args struct {
-		revisions []servingv1.Revision
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    int
-		wantErr bool
-	}{
-		{
-			name: "should return highest generation from label",
-			args: args{
-				revisions: []servingv1.Revision{
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "1"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "2"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "3"}}},
-				},
-			},
-			want: 3,
-		},
-		{
-			name: "should return error if even one revision lacks proper label",
-			args: args{
-				revisions: []servingv1.Revision{
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "1"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "2"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{"random-label": "3"}}},
-				},
-			},
-			want:    -1,
-			wantErr: true,
-		},
-		{
-			name: "should return error generation is not a number",
-			args: args{
-				revisions: []servingv1.Revision{
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "1"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "NaN"}}},
-				},
-			},
-			want:    -1,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := gomega.NewGomegaWithT(t)
-
-			got, err := getNewestGeneration(tt.args.revisions)
-			g.Expect(err != nil).To(gomega.Equal(tt.wantErr))
-			g.Expect(got).To(gomega.Equal(tt.want))
-		})
-	}
-}
-
-func TestFunctionReconciler_getOldRevisionSelector(t *testing.T) {
-	// serverless.kyma-project.io/uuid=uid,serving.knative.dev/configurationGeneration!=3
-	type args struct {
-		instance  *serverlessv1alpha1.Function
-		revisions []servingv1.Revision
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    labels.Selector
-		wantErr bool
-	}{
-		{
-			name: "properly parses revisions and instances",
-			args: args{
-				instance: &serverlessv1alpha1.Function{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "uid",
-					},
-				},
-				revisions: []servingv1.Revision{
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "1"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "2"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "3"}}},
-				},
-			},
-			wantErr: false,
-			want: func() labels.Selector {
-				lbl, _ := labels.Parse("serverless.kyma-project.io/uuid=uid,serving.knative.dev/configurationGeneration!=3")
-				return lbl
-			}(),
-		},
-		{
-			name: "fails with incorrect labels from revisions",
-			args: args{
-				instance: &serverlessv1alpha1.Function{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "uid",
-					},
-				},
-				revisions: []servingv1.Revision{
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "1"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "2"}}},
-					{ObjectMeta: metav1.ObjectMeta{Labels: map[string]string{cfgGenerationLabel: "ups"}}},
-				},
-			},
-			wantErr: true,
-			want:    nil,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := gomega.NewGomegaWithT(t)
-
-			r := &FunctionReconciler{}
-			got, err := r.getOldRevisionSelector(tt.args.instance, tt.args.revisions)
-
-			g.Expect(err != nil).To(gomega.Equal(tt.wantErr))
-			if got != nil {
-				g.Expect(got).To(gomega.Equal(tt.want))
-			} else {
-				g.Expect(got).To(gomega.BeNil())
-			}
-
-		})
-	}
-}
